Drop musllinux builds (pyproj isn't providing them)
