+++ pyresample/test/test_geometry/test_area.py
@@ -1849,16 +1849,22 @@ class TestAreaDefGetAreaSlices:
         crop_area = create_test_area({'proj': 'latlong'},
                                      100, 100,
                                      (15.9689, 58.5284, 16.4346, 58.6995))
+        cache_glob = str(tmp_path / "geometry_slices_v1" / "*.json")
         with pyresample.config.set(cache_dir=tmp_path, cache_geom_slices=False):
-            assert len(glob(str(tmp_path / "geometry_slices_v1" / "*.json"))) == 0
+            assert len(glob(cache_glob)) == 0
             slice_x, slice_y = src_area.get_area_slices(crop_area)
-            assert len(glob(str(tmp_path / "geometry_slices_v1" / "*.json"))) == 0
+            assert len(glob(cache_glob)) == 0
         with pyresample.config.set(cache_dir=tmp_path, cache_geom_slices=True):
-            assert len(glob(str(tmp_path / "geometry_slices_v1" / "*.json"))) == 0
+            assert len(glob(cache_glob)) == 0
             slice_x, slice_y = src_area.get_area_slices(crop_area)
-            assert len(glob(str(tmp_path / "geometry_slices_v1" / "*.json"))) == 1
+            assert len(glob(cache_glob)) == 1
         assert slice_x == slice(5630, 8339)
         assert slice_y == slice(9261, 10980)
+
+        from pyresample.future.geometry._subset import get_area_slices
+        with pyresample.config.set(cache_dir=tmp_path):
+            get_area_slices.cache_clear()
+        assert len(glob(cache_glob)) == 0
 
 
 class TestBoundary:
